[Refactor] 예외처리 추가

+++ src/main/java/com/example/board/bookmark/repository/BookmarkRepository.java
@@ -1,5 +1,7 @@
 package com.example.board.bookmark.repository;
 
+import static com.example.board.global.exception.GlobalExceptionCode.NOT_FOUND_BOOKMARK;
+
 import com.example.board.bookmark.entity.Bookmark;
 import com.example.board.member.entity.Member;
 import java.util.List;
@@ -16,18 +18,24 @@ public class BookmarkRepository {
     bookmarkJpaRepository.save(bookmark);
   }
 
-  public Bookmark findByBoardIdAndMember(Long boardId, Member member) {
-    return bookmarkJpaRepository.findByBoardIdAndMember(boardId, member);
+  public Bookmark findByBoardId(Long boardId) {
+    return bookmarkJpaRepository.findByBoardId(boardId).orElseThrow(NOT_FOUND_BOOKMARK::newException);
   }
 
   public void delete(Long boardId, Member member) {
-    Bookmark bookmark = findByBoardIdAndMember(boardId,member);
+    Bookmark bookmark = findByBoardId(boardId);
+
+    bookmark.validateAccess(member);
 
     bookmarkJpaRepository.delete(bookmark);
   }
 
   public List<Long> findBoardIdByMember(Member member) {
     return bookmarkJpaRepository.findBoardIdByMember(member);
+  }
+
+  public boolean existsByBoardIdAndMember(Long boardId, Member member) {
+    return bookmarkJpaRepository.existsByBoardIdAndMember(boardId, member);
   }
 
 }
